+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.10
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.7.9
+
 ## 3.0.9
 
 ### Patch Changes